# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservations - API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cinemas/{cinemaId}/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Gets the list of rooms by cinema",
                "description": "Returns the list of rooms by cinema with optional pagination",
                "operationId": "getRoomsListByCinema",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of cinemas per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "cinemaId",
                        "in": "path",
                        "description": "ID of cinema to return rooms",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No room(s) found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Store a new room by cinema",
                "description": "Stores a new room by cinema and returns the room data",
                "operationId": "storeRoomByCinema",
                "parameters": [
                    {
                        "name": "cinemaId",
                        "in": "path",
                        "description": "ID of cinema to store room",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Room data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "seats"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    },
                                    "seats": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "seats": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "cinema_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/cinemas/{cinemaId}/rooms/{roomId}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get room by ID by cinema",
                "description": "Returns a single room by cinema",
                "operationId": "getRoomByIdByCinema",
                "parameters": [
                    {
                        "name": "cinemaId",
                        "in": "path",
                        "description": "ID of cinema to return",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "ID of room to return",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "seats": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "cinema_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update an existing room by cinema",
                "description": "Updates and returns a room by cinema data",
                "operationId": "updateRoomByCinema",
                "parameters": [
                    {
                        "name": "cinemaId",
                        "in": "path",
                        "description": "ID of cinema room that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "ID of room that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Room data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "seats"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    },
                                    "seats": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "seats": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "cinema_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room by cinema",
                "description": "Deletes a room by cinema and returns a success message",
                "operationId": "deleteRoomByCinema",
                "parameters": [
                    {
                        "name": "cinemaId",
                        "in": "path",
                        "description": "ID of cinema to delete room",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "ID of room to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room deleted successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "500": {
                        "description": "Error deleting room"
                    }
                }
            }
        },
...
    }
}